Insertion of stats into db

+++ Industrial_Informatics_Project/DataModel/Quiz_Stats.cs
@@ -18,7 +18,7 @@ namespace Industrial_Informatics_Project.DataModel
         public Nullable<int> user_id { get; set; }
         public Nullable<int> corect_answers { get; set; }
         public Nullable<int> incorect_answers { get; set; }
-        public Nullable<float> time_remaining { get; set; }
+        public Nullable<double> time_remaining { get; set; }
         public Nullable<System.DateTime> date { get; set; }
         public Nullable<int> difficulty { get; set; }
     }
